- IconGrid now sorts all items it contains. - Cleaned up some warnings in PlanetSelectState. - Structure now defines a Storage Pool. - MineFacility and Agridome now take advantage of the storage pool interface in Structure. Closes IssueID #37.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@52 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/States/PlanetSelectState.cpp
@@ -37,9 +37,9 @@ void PlanetSelectState::initialize()
 	mPlanets.push_back(new Planet(PLANET_TYPE_MARS));
 	mPlanets.push_back(new Planet(PLANET_TYPE_GANYMEDE));
 
-	mPlanets[0]->position(r.width() / 4 - 64, r.height() / 2 - 64);
-	mPlanets[1]->position(r.width() / 2 - 64, r.height() / 2 - 64);
-	mPlanets[2]->position(((r.width() / 4) * 3) - 64, r.height() / 2 - 64);
+	mPlanets[0]->position((int)r.width() / 4 - 64, (int)r.height() / 2 - 64);
+	mPlanets[1]->position((int)r.width() / 2 - 64, (int)r.height() / 2 - 64);
+	mPlanets[2]->position((((int)r.width() / 4) * 3) - 64, (int)r.height() / 2 - 64);
 
 	PLANET_TYPE_SELECTION = PLANET_TYPE_NONE;
